feat: integrate mermaid in markdown code-blocks
